Move some code into a new file platform.hpp/.cpp
